Relations: @ManyToOne

+++ src/main/java/lab/lhss/ecommerce/model/Client.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.List;
 
 @Entity
 @Getter
@@ -16,7 +17,10 @@ public class Client {
     @Id
     @EqualsAndHashCode.Include
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
+    //@OneToMany(mappedBy = "client")
+    //private List<Order> orders;
 
     private String name;
